Fix smart_money_congress id NOT auto-incrementing (SQLite schema bug)
SQLite only auto-increments a PRIMARY KEY declared as the exact type
INTEGER — BigInteger maps to BIGINT which does not alias the rowid,
causing every ORM insert to fail with NOT NULL constraint on id.

- smart_money.py: Integer (not BigInteger) for id column
- m013_fix_smart_money_congress_id.py: idempotent migration that checks
  PRAGMA table_info, drops the empty table, recreates with INTEGER PRIMARY
  KEY AUTOINCREMENT, and rebuilds both indices
- main.py: wire up m013 in the startup migration chain
- congress.py: per-row flush + IntegrityError catch so a single duplicate
  can't roll back the entire 200+ row batch; fetch error now also returns
  early without losing already-counted rows

After deploy: next startup runs m013 → table recreated → startup seed
fires → Fetch Now inserts all rows with no errors.

+++ backend/app/db/models/smart_money.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from datetime import date, datetime, timezone
 
-from sqlalchemy import BigInteger, Column, Date, DateTime, Index, String, Text, UniqueConstraint
+from sqlalchemy import BigInteger, Column, Date, DateTime, Index, Integer, String, Text, UniqueConstraint
 from sqlalchemy.sql import func
 
 from app.db.base import Base
@@ -11,7 +11,8 @@ from app.db.base import Base
 class SmartMoneyCongressTrade(Base):
     __tablename__ = "smart_money_congress"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    # Integer (not BigInteger) required for SQLite rowid alias → auto-increment
+    id = Column(Integer, primary_key=True, autoincrement=True)
     member_name = Column(String(200), nullable=False)
     party = Column(String(1))
     chamber = Column(String(1))           # 'S' or 'H'
